feat: Generat sample docs, move validate outside
When a user uploads their own policy document, it is not clear for them how to create documents to validate against the extracted rules in #validate-file-input.

So, add a "Generate sample documents" next to the green "Validate" button that, when clicked, uses an LLM call similar to the other calls (showing a loading indicator) and generates 2 documents using the context from the current rule set: one that will likely pass and one that will fail. Display 2 buttons below the validate button allowing users to click on each to download.

Also, move the Storage - Clear all button outside of the advanced settings into the same row as Ingest - Consolidate - Clear all.